refactor: move function definitions to db_populator module
